+++ x/delayedack/types/packet_metadata.go
@@ -1,6 +1,9 @@
 package types
 
 import (
+	"encoding/json"
+	"fmt"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -22,4 +25,38 @@ func (e EIBCMetadata) ValidateBasic() error {
 		return ErrInvalidEIBCFee
 	}
 	return nil
+}
+
+const (
+	memoObjectKeyEIBC = "eibc"
+	memoObjectKeyPFM  = "forward"
+)
+
+var (
+	ErrMemoUnmarshal = fmt.Errorf("unmarshal memo")
+	ErrMemoIsPFM     = fmt.Errorf("EIBC packet with PFM is currently not supported")
+	ErrMemoEibcEmpty = fmt.Errorf("memo IBC field is missing")
+)
+
+func ParsePacketMetadata(input string) (*PacketMetadata, error) {
+	bz := []byte(input)
+
+	memo := make(map[string]any)
+	err := json.Unmarshal(bz, &memo)
+	if err != nil {
+		return nil, ErrMemoUnmarshal
+	}
+	if memo[memoObjectKeyPFM] != nil {
+		// Currently not supporting eibc with PFM: https://github.com/dymensionxyz/dymension/issues/599
+		return nil, ErrMemoIsPFM
+	}
+	if memo[memoObjectKeyEIBC] == nil {
+		return nil, ErrMemoEibcEmpty
+	}
+	var metadata PacketMetadata
+	err = json.Unmarshal(bz, &metadata)
+	if err != nil {
+		return nil, ErrMemoUnmarshal
+	}
+	return &metadata, nil
 }
